Added --install-only and renamed constants package
