+++ lint_python/__init__.py
@@ -1,3 +1,3 @@
-from .__version__ import __version__
+from .constants import __version__
 
 __all__ = ["__version__"]
