--- lint_python/__version__.py
@@ -1,2 +1,0 @@
-__version__ = "2.0.0"
-USED_LINTERS = ["isort==5.10.1", "black==22.3.0", "flake8==3.8.4", "mypy==0.940"]
