+++ lint_python/constants.py
@@ -1,0 +1,2 @@
+__version__ = "2.0.0"
+USED_LINTERS = ["isort==5.10.1", "black==22.6.0", "flake8==4.0.1", "mypy==0.971"]
